Updated 20180618-0128

+++ preparation/parse_xml.py
@@ -8,7 +8,8 @@ import re
 from collections import defaultdict
 import configparser
 import shutil
-
+import gzip
+import datetime
 # ### Using XPATH
 
 
@@ -17,6 +18,12 @@ def get_app_ref_doc_number():
     for path in root.findall("./bibliographic-data/application-reference/document-id/doc-number"):
         app_ref_doc_num = path.text
     return app_ref_doc_num
+
+######################## application_date
+def get_application_date():
+    for path in root.findall("./bibliographic-data/application-reference/document-id/date"):
+        application_date = path.text
+    return application_date
 
 
 ######################## classification-ipcr
@@ -44,10 +51,6 @@ def get_classification_ipcr():
     return (classification_text, label_lookup_key_list)
 
 
-t = 'test'
-t.ljust(10, '0')
-
-
 ######################## invention-title in English
 
 def get_invention_title():
@@ -56,6 +59,11 @@ def get_invention_title():
         lang = path.get("lang")
         if lang == "EN":
             inv_title = path.text
+  
+    # TODO: Limit to top 150 words instead of first 150
+    # Truncating to 150 words
+    inv_title = " ".join( inv_title.split()[:limit_word_count] )
+    
     return inv_title
 
 
@@ -72,6 +80,11 @@ def get_abstract_text():
                 abstract_text = abstract_text + " " + (abstract_text_tmp or '')
             abstract_text = re.sub(pattern, '', abstract_text)
             abstract_text = re.sub('\s+', ' ', abstract_text.replace('\\n', ' ')).strip()
+  
+    # TODO: Limit to top 150 words instead of first 150
+    # Truncating to 150 words
+    abstract_text = " ".join( abstract_text.split()[:limit_word_count] )
+    
     return abstract_text
 
 
@@ -88,6 +101,10 @@ def get_description_text():
                 desc_text = desc_text + " " + (child.text or '')
             # replacing newlines and multiple spaces
             desc_text = re.sub('\s+', ' ', desc_text.replace('\\n', ' ')).strip()
+  
+    # TODO: Limit to top 150 words instead of first 150
+    # Truncating to 150 words
+    desc_text = " ".join( desc_text.split()[:limit_word_count] )
 
     return desc_text
 
@@ -108,6 +125,10 @@ def get_claim_text():
                 claim_text = claim_text + " " + (claim_text_tmp or '')
             claim_text = re.sub(pattern, '', claim_text)
             claim_text = re.sub('\s+', ' ', claim_text.replace('\\n', ' ')).strip()
+ 
+    # TODO: Limit to top 150 words instead of first 150
+    # Truncating to 150 words
+    claim_text = " ".join( claim_text.split()[:limit_word_count] )
     return claim_text
 
 
@@ -160,22 +181,55 @@ def read_label_files():
 
 ######################## Get the labels
 
-def get_labels():
+def get_labels( l_file_path_new,  l_app_ref_doc_num):
     labels_list = []
     for key in label_lookup_key_list:
+        if key not in label_dict.keys():
+            # Checks for missing label keys missing in label source and writes them to error files
+            error_message=("LabelKey:{key}|File:{l_file_path_new}|AppRefDocNumber:{l_app_ref_doc_num}"
+            .format(
+                    key=key,
+                    l_file_path_new=l_file_path_new,
+                    l_app_ref_doc_num=l_app_ref_doc_num))
+            
+            with open(join(errorDIR, fileMissingLabel), "a+") as f:
+                f.write(error_message + '\n')
+
+            continue
         labels_list.append(label_dict[key])
     labels_string = "|".join(str(e) for e in labels_list)
     return labels_string
 
 ######################## Remove existing files
-def remove_files():
-    for filename in os.listdir(outDIR):
-        filepath = os.path.join(outDIR, filename)
-        try:
-            shutil.rmtree(filepath)
-        except OSError:
-            os.remove(filepath)
-    
+def remove_files(dir_list=[]):
+    new_dir_list = [ dir_name+"_zipped" for dir_name in dir_list]
+    new_dir_list.extend(dir_list)
+    for dir_name in new_dir_list:
+        print("remove_files: working on {}".format(dir_name))
+        if os.path.exists(dir_name) is False:
+            continue
+
+        for filename in os.listdir(dir_name):
+            filepath = os.path.join(dir_name, filename)
+            try:
+                shutil.rmtree(filepath)
+            except OSError:
+                os.remove(filepath)
+
+######################## Zip output files
+def zip_output(dir_list=[]):
+    for dir_name in dir_list:
+        zipDIR = dir_name+"_zipped"
+
+        if os.path.exists(zipDIR) is False:
+            os.makedirs(zipDIR)
+        for root, directories, files in os.walk(dir_name):
+            for filename in files:
+                filepath = os.path.join(root, filename)
+                with open(filepath, 'rb') as f_in:
+                    with gzip.open(os.path.join(zipDIR, filename+".gz"), 'wb') as f_out:
+                        shutil.copyfileobj(f_in, f_out)
+
 ######################## MAIN ########################
 
 config = configparser.ConfigParser()
@@ -184,6 +238,11 @@ config.read('../config/preparation.ini')
 inDIR = config['DEFAULT']['inDIR']
 pattern = config['DEFAULT']['pattern']
 outDIR = config['DEFAULT']['outDIR']
+errorDIR = config['DEFAULT']['errorDIR']
+labelDIR = config['DEFAULT']['labelDIR']
+logDIR = config['DEFAULT']['logDIR']
+
+
 fileNameToDocNumber = config['DEFAULT']['fileNameToDocNumber']
 fileDocNumberToClassText = config['DEFAULT']['fileDocNumberToClassText']
 fileDocNumberToInvTitle = config['DEFAULT']['fileDocNumberToInvTitle']
@@ -191,15 +250,24 @@ fileDocNumberToAbsText = config['DEFAULT']['fileDocNumberToAbsText']
 fileDocNumberToDescText = config['DEFAULT']['fileDocNumberToDescText']
 fileDocNumberToClaimText = config['DEFAULT']['fileDocNumberToClaimText']
 fileDocNumberToLabels = config['DEFAULT']['fileDocNumberToLabels']
+fileMissingLabel = config['DEFAULT']['fileMissingLabel']
+
+
 labelPattern = config['DEFAULT']['labelPattern']
-labelDIR = config['DEFAULT']['labelDIR']
 limit_files_write = config.getint('DEFAULT','limit_files_write')
 path_string_to_replace = config['DEFAULT']['path_string_to_replace']
-
+limit_word_count = config.getint('DEFAULT','limit_word_count')
 fileList = []
 doc_number_filename = defaultdict(list)
 cnt_files = 0
 
+log_file = datetime.datetime.now().strftime("%Y%m%d-%H%M%S")+".log"
+
+
+print()
+# with open(join(logDIR, fileDocNumberToDescText), "a+") as f:
+#     write_str = app_ref_doc_num + "|" + desc_text
+#     f.write(write_str + '\n')
 
 # Read label files
 label_dict = {}
@@ -208,16 +276,18 @@ label_dict = read_label_files()
 # For brevity in output file fileNameToDocNumber.txt
 pattern_path = re.compile("^"+path_string_to_replace)
 
-#Removing existing files
+#Remove existing files. The logic will also consider *_zipped directories
 
-remove_files()
+dir_to_delete_list=[outDIR,errorDIR]
+remove_files(dir_to_delete_list)
 
+
+docNumber_date_dict={}
 for dName, sdName, fList in os.walk(inDIR):
     outer_break_flag = 0
     for fileName in fList:
         # Match search pattern
         if fnmatch.fnmatch(fileName, pattern):
-            cnt_files = cnt_files + 1
             
             # Controls the number of files read
             if cnt_files == limit_files_write:
@@ -230,21 +300,49 @@ for dName, sdName, fList in os.walk(inDIR):
             tree = ET.parse(file_path)
             root = tree.getroot()
             app_ref_doc_num = get_app_ref_doc_number()
+            application_date=get_application_date()
 
+            # Ignoring if patent with same application number has been seen
+            if app_ref_doc_num in docNumber_date_dict.keys():
+                existing_patent_date=docNumber_date_dict[app_ref_doc_num]
+                with open(join(logDIR, log_file), "a+") as f:
+                    write_str = ("doc-number {}, application_date {} has been seen with date {}".format(app_ref_doc_num,existing_patent_date,application_date))
+                    f.write(write_str + '\n')
+                continue
+
+            docNumber_date_dict[app_ref_doc_num] = application_date
+                
             label_lookup_key_list = []
             (classification_text, label_lookup_key_list) = get_classification_ipcr()
-            labels = get_labels()
+            labels = get_labels(file_path_new,app_ref_doc_num)
+ 
             inv_title = get_invention_title()
+            if inv_title == 'NA':
+                continue
+
             abstract_text = get_abstract_text()
+            if abstract_text == 'NA':
+                continue
+
             desc_text = get_description_text()
+            if desc_text == 'NA':
+                continue
+
             claim_text = get_claim_text()
+            if claim_text == 'NA':
+                continue
+
+            cnt_files = cnt_files + 1
 
             write_files()
+            
 
     if outer_break_flag == 1:
         break
 
-
+# Compress output for sharing. The logic will also consider *_zipped directories
+dir_to_zip_list=[outDIR,errorDIR,logDIR]
+zip_output(dir_to_zip_list)
